// app/routes/interfaces/Product.ts
export interface Product {
    produto: {
      id: number;
      codigo: string;
      descricao: string;
      tipo: string;
      situacao: string;
      unidade: string;
      preco: string;
      precoCusto: string;
      descricaoCurta: string;
      descricaoComplementar: string;
      dataInclusao: string;
      dataAlteracao: string;
      imageThumbnail: string;
      urlVideo: string;
      nomeFornecedor: string;
      codigoFabricante: string;
      marca: string;
      class_fiscal: string;
      cest: string;
      origem: string;
      idGrupoProduto: string;
      linkExterno: string;
      observacoes: string;
      grupoProduto: string;
      garantia: string;
      descricaoFornecedor: string;
      idFabricante: string;
      pesoLiq: string;
      pesoBruto: string;
      estoqueMinimo: string;
      estoqueMaximo: string;
      gtin: string;
      gtinEmbalagem: string;
      larguraProduto: string;
      alturaProduto: string;
      profundidadeProduto: string;
      unidadeMedida: string;
      itensPorCaixa: string;
      volumes: string;
      localizacao: string;
      crossdocking: string;
      condicao: string;
      freteGratis: string;
      producao: string;
      dataValidade: string;
      spedTipoItem: string;
    };
  }

 

  